refactor: unify keyboard_capture and modifier_monitor into keyboard_service
Replace the separate keyboard_capture (device_query polling for shortcut
recording) and modifier_monitor (standalone modifier key shortcuts) with
a single keyboard_service that manages one polling thread with modal
capture/monitoring states. Entering capture mode is now a single IPC
call that atomically switches mode, unregisters shortcuts, and starts
reporting key presses. Exiting re-registers everything from config.

- Add capture-mode guard in shortcut manager to discard phantom triggers
- Simplify ShortcutInput: remove manual pause/resume of global shortcuts
- Clean up dead code in tray.rs (duplicate is_dark_theme, unused import)
